Rename handlers folder to http
