+++ http/handlers/projects/create.go
@@ -3,8 +3,8 @@ package projects
 import (
 	"net/http"
 
-	"github.com/4lexvav/hit-the-goal/handlers/common"
-	"github.com/4lexvav/hit-the-goal/handlers/projects/requests"
+	"github.com/4lexvav/hit-the-goal/http/common"
+	"github.com/4lexvav/hit-the-goal/http/handlers/projects/requests"
 	"github.com/4lexvav/hit-the-goal/services"
 )
